perf(logo): add lazy loading
