+++ packages/components/src/components/bal-logo/bal-logo.small.tsx
@@ -1,0 +1,12 @@
+import { h } from '@stencil/core'
+
+export const logo = (
+  <svg width="100" height="22" viewBox="0 0 100 22" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path
+      fill-rule="evenodd"
+      clip-rule="evenodd"
+      d="M71.2886 3.08227C71.2886 1.78086 72.3924 0.84771 73.679 0.84771C74.9656 0.84771 75.9688 1.78086 75.9688 3.08227C75.9688 4.38368 74.9656 5.29533 73.679 5.29533C72.3924 5.29533 71.2886 4.38403 71.2886 3.08227ZM71.8472 6.83532H75.4888V18.2682H71.8472V6.83532ZM52.405 2.40989H56.0469V18.2678H52.405V2.40989ZM30.6826 18.5505C34.0192 18.5505 36.3961 15.9903 36.3961 12.5629C36.3961 9.13507 34.0847 6.57525 30.7687 6.5756C29.5268 6.5756 28.48 6.94421 27.6514 7.63843V2.41025H24.0095V18.2682H27.368V17.292C28.1966 18.0948 29.3089 18.5505 30.6826 18.5505ZM30.1594 9.74366C31.6858 9.74366 32.7978 10.9358 32.7978 12.5629C32.7978 14.1899 31.6858 15.3613 30.1594 15.3613C28.5675 15.3613 27.4768 14.2114 27.4768 12.5629C27.4768 10.8932 28.5455 9.74366 30.1594 9.74366ZM50.0036 18.2682V6.83532H46.6451V8.02854C45.8382 7.22613 44.7475 6.79197 43.3305 6.79197C39.994 6.79197 37.5951 9.13543 37.5951 12.5629C37.5951 15.9903 39.994 18.5505 43.3305 18.5505C44.7478 18.5505 45.8382 18.0726 46.6451 17.2269V18.2682H50.0036ZM43.8541 9.74367C45.446 9.74367 46.5147 10.8921 46.5147 12.5629C46.5147 14.2117 45.4457 15.3613 43.8541 15.3613C42.3061 15.3613 41.2154 14.1888 41.2154 12.5629C41.2154 10.9369 42.3277 9.74367 43.8541 9.74367ZM63.9216 6.57524C60.2581 6.57524 57.641 9.13507 57.641 12.5628C57.641 15.9688 60.2591 18.5505 63.9216 18.5505C67.6071 18.5505 70.2238 15.9688 70.2238 12.5628C70.2238 9.13507 67.6057 6.57524 63.9216 6.57524ZM63.9432 9.78595C65.5351 9.78595 66.5603 10.914 66.5603 12.541C66.5603 14.1899 65.5135 15.3394 63.9432 15.3398C62.3513 15.3398 61.3045 14.168 61.3045 12.5628C61.3045 10.9136 62.3513 9.78595 63.9432 9.78595ZM79.9809 14.3418L76.9717 15.4916C77.6475 17.489 79.5013 18.5505 82.095 18.5484C85.1262 18.5484 87.045 17.1166 87.045 14.7516C87.045 12.1567 84.765 11.5576 83.0168 11.0983C81.9297 10.8126 81.0482 10.581 81.0482 9.95722C81.0482 9.48007 81.4187 9.19815 82.1601 9.19815C82.9235 9.19815 83.4687 9.56676 83.7305 10.2828L86.7415 9.11111C86.2834 7.50664 84.5824 6.57384 82.2271 6.57384C79.327 6.57384 77.4949 7.92035 77.4949 10.1331C77.4949 12.4741 79.6268 13.1109 81.3484 13.6252C82.5143 13.9735 83.4921 14.2655 83.4921 14.9926C83.4921 15.5131 83.056 15.8384 82.2271 15.8384C81.0496 15.8384 80.2862 15.2964 79.9809 14.3418ZM99.9999 12.3243C99.9999 9.50508 97.7756 6.5534 94.0246 6.5534C90.7101 6.5534 88.0494 9.26686 88.0494 12.5628C88.0494 15.9688 90.4483 18.5505 94.2429 18.5505C96.7469 18.5505 98.2259 17.5905 99.1125 16.496L97.0183 14.4084C96.3644 15.2132 95.6385 15.5998 94.3516 15.5998C92.9779 15.5998 91.9531 14.9056 91.5822 13.6264H99.9128C99.9728 13.195 100.002 12.7598 99.9999 12.3243ZM91.5383 11.3696C91.8436 9.98118 92.7813 9.2651 94.1114 9.2651C95.4416 9.2651 96.3141 10.0679 96.6194 11.3696H91.5383ZM10.6109 1.49718L19.7559 10.5947C19.8357 10.6742 19.899 10.7686 19.9422 10.8725C19.9854 10.9764 20.0076 11.0878 20.0075 11.2002C20.0074 11.3126 19.985 11.4239 19.9416 11.5277C19.8983 11.6315 19.8348 11.7258 19.7548 11.8052L17.2751 14.272L10.003 7.03759L5.81949 11.1994L10.003 15.3613L12.5967 12.7806L15.6857 15.8532L10.6127 20.9017C10.5328 20.9811 10.438 21.0442 10.3337 21.0872C10.2294 21.1302 10.1175 21.1523 10.0046 21.1523C9.89169 21.1523 9.77986 21.1302 9.67553 21.0872C9.5712 21.0442 9.47641 20.9811 9.39657 20.9017L0.251611 11.8041C0.0904986 11.6437 0 11.4262 0 11.1994C0 10.9726 0.0904986 10.7551 0.251611 10.5947L9.3948 1.49718C9.47464 1.41772 9.56943 1.35469 9.67376 1.31169C9.77809 1.26868 9.88992 1.24655 10.0028 1.24655C10.1158 1.24655 10.2276 1.26868 10.3319 1.31169C10.4363 1.35469 10.5311 1.41772 10.6109 1.49718Z"
+      fill="#000D6E"
+    />
+  </svg>
+)
